requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.python.org/simple
appdirs==1.4.4
astropy==4.2
certifi==2020.12.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cycler==0.10.0
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
joblib==1.0.1
kiwisolver==1.3.1; python_version >= '3.6'
matplotlib==3.3.4; python_version >= '3.6'
numpy==1.20.1
packaging==20.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pandas==1.2.3
pillow==8.1.2; python_version >= '3.6'
pooch==1.3.0; python_version >= '3.6'
pyerfa==1.7.2; python_version >= '3.6'
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
pyproj==3.0.1
pyshtools==4.8
pytest-runner==5.3.0
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
pytz==2021.1
requests==2.25.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
scipy==1.6.2
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
tqdm==4.59.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
xarray==0.17.0; python_version >= '3.7'
